change utilities to module and add first snapping functions
