+++ runtime-rust/src/utilities/noise.rs
@@ -1,0 +1,56 @@
+use openssl::rand::rand_bytes;
+use byteorder::{ByteOrder, LittleEndian};
+use probability::distribution::{Gaussian, Laplace, Inverse};
+use ieee754::Ieee754;
+
+use crate::utilities::snapping;
+
+pub fn sample_laplace(shift: f64, scale: f64) -> f64 {
+    let probability: f64 = sample_uniform(0., 1.);
+    Laplace::new(shift, scale).inverse(probability)
+//    shift - scale * (sample - 0.5).signum() * (1. - 2. * (sample - 0.5).abs()).ln()
+}
+
+pub fn sample_gaussian(shift: f64, scale: f64) -> f64 {
+    let probability: f64 = sample_uniform(0., 1.);
+    Gaussian::new(shift, scale).inverse(probability)
+}
+
+pub fn sample_uniform(min: f64, max: f64) -> f64 {
+    let mut buf: [u8; 8] = [0; 8];
+    rand_bytes(&mut buf).unwrap();
+    (LittleEndian::read_u64(&buf) as f64) / (std::u64::MAX as f64) * (max - min) + min
+}
+
+pub fn sample_uniform_snapping() -> f64 {
+    /// Returns random sample from Uniform(0,1)
+    ///
+    /// This algorithm is taken from Mironov (2012) http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.366.5957&rep=rep1&type=pdf
+    /// and is important for making some of the guarantees in the paper.
+    ///
+    /// The idea behind the uniform sampling is to first sample a "precision band".
+    /// Each band is a range of floating point numbers with the same level of arithmetic precision
+    /// and is situated between powers of two.
+    /// A band is sampled with probability relative to the unit of least precision using the Geometric distribution.
+    /// That is, the uniform sampler will generate the band [1/2,1) with probability 1/2, [1/4,1/2) with probability 1/4,
+    /// and so on.
+    ///
+    /// Once the precision band has been selected, floating numbers numbers are generated uniformly within the band
+    /// by generating a 52-bit mantissa uniformly at random.
+
+    // Generate mantissa
+    let binary_string = snapping::get_bytes(7);
+    let mantissa = &binary_string[0..52];
+
+    // convert mantissa to integer
+    let mantissa_int = u64::from_str_radix(mantissa, 2).unwrap();
+
+    // Generate exponent
+    let geom: (i16) = snapping::get_geom_prob_one_half();
+    let exponent: (u16) = (-geom + 1023) as u16;
+
+    // Generate uniform random number from (0,1)
+    let uniform_rand = f64::recompose_raw(false, exponent, mantissa_int);
+
+    return uniform_rand;
+}
